Check if tmp folder exists and exit code handling
The cpatureScreenshot() method now checks if the /tmp folder exists and creates it if not. This happens as a first action upon calling.

Also, some exit code handling for the screencapture task was added. For now, the method returns 0 if the user aborts the screencapture action. This may change later according to the data the method returns then.